check before backup old files

+++ deploy.sh
@@ -2,6 +2,6 @@
 
 for i in zshenv zshrc zshrc.aliases zshrc.funcs
 do
-    cp -v ~/.$i ~/.$i.bkp
+    [ ! -f ~/.$i ] || cp -v ~/.$i ~/.$i.bkp
     cp -v dot_$i ~/.$i
 done
